// util/interfaces.ts
export interface IMatchOptions {
    beginTime?: number;
    endTime?: number;
}
export interface IAPIMatchListWrapper {
    matches: IAPIMatchList[];
    startIndex: number;
    endIndex: number;
    totalGames: number;
}
export interface IAPIMatchList {
    platformId: string;
    gameId: number;
    champion: number;
    queue: number;
    season: number;
    timestamp: number;
    role: string;
    lane: string;
}
export interface IDBMatchList {
    id?: number;
    summonerPUUID: string;
    gameId: number;
    championId: number;
    lane: string;
    role: string;
    timestamp: Date;
}
export interface IDBDeltaTypes {
    id: number;
    name: string;
}
export interface IDBMatch {
    gameId: number;
    seasonId: number;
    queueId: number;
    mapId: number;
    platformId: string;
    gameVersion: string;
    gameMode: string;
    gameType: string;
    gameDuration: number;
    gameCreation: Date;
}
// ...
export interface IAPIBan {
    championId: number;
    pickTurn: number;
}
export interface IAPIParticipant {
    championId: number;
    highestAchievedSeasonTier?: string;
    participantId: number;
    spell1Id: number;
    spell2Id: number;
    stats: IAPIParticipantStats;
    teamId: number;
    timeline: IParticipantTimeline;
}
export interface IDBParticipant {
    accountId: string;
    championId: number;
    gameId: number;
    highestAchievedSeasonTier: string;
    id?: number;
    lane: string;
    participantId: number;
    role: string;
    spell1Id: number;
    spell2Id: number;
    teamId: number;
}
// ...
export interface IDBParticipantTimelineDelta {
    id?: number;
    participantId: number;
    deltaTypeId: number;
    increment: string;
    value: number;
}
// ...
export interface IDBTeamStat {
    id?: number;
    gameId: number;
    teamId: number;
    win: string;
    baronKills: number;
    riftHeraldKills: number;
    vilemawKills: number;
    inhibitorKills: number;
    towerKills: number;
    dragonKills: number;
    dominionVictoryScore: number;
    firstDragon: boolean;
    firstInhibitor: boolean;
    firstRiftHerald: boolean;
    firstBlood: boolean;
    firstTower: boolean;
}
export interface IDBTeamBan extends IAPIBan {
    gameId: number;
    id?: number;
    teamId: number;
}
export interface IDBParticipantStats {
    id?: number;
    neutralMinionsKilledTeamJungle: number;
    visionScore: number;
    magicDamageDealtToChampions: number;
    largestMultiKill: number;
    totalTimeCrowdControlDealt: number;
    longestTimeSpentLiving: number;
    tripleKills: number;
    kills: number;
    totalScoreRank: number;
    neutralMinionsKilled: number;
    damageDealtToTurrets: number;
    physicalDamageDealtToChampions: number;
    damageDealtToObjectives: number;
    totalUnitsHealed: number;
    totalDamageTaken: number;
    wardsKilled: number;
    largestCriticalStrike: number;
    largestKillingSpree: number;
    quadraKills: number;
    magicDamageDealt: number;
    firstBloodAssist: boolean;
    damageSelfMitigated: number;
    magicalDamageTaken: number;
    firstInhibitorKill: boolean;
    trueDamageTaken: number;
    assists: number;
    goldSpent: number;
    trueDamageDealt: number;
    participantId: number;
    physicalDamageDealt: number;
    sightWardsBoughtInGame: number;
    totalDamageDealtToChampions: number;
    physicalDamageTaken: number;
    totalPlayerScore: number;
    win: boolean;
    objectivePlayerScore: number;
    totalDamageDealt: number;
    neutralMinionsKilledEnemyJungle: number;
    deaths: number;
    wardsPlaced: number;
    perkPrimaryStyle: number;
    perkSubStyle: number;
    turretKills: number;
    firstBloodKill: boolean;
    trueDamageDealtToChampions: number;
    goldEarned: number;
    killingSprees: number;
    unrealKills: number;
    firstTowerAssist: boolean;
    firstTowerKill: boolean;
    champLevel: number;
    doubleKills: number;
    inhibitorKills: number;
    firstInhibitorAssist: boolean;
    combatPlayerScore: number;
    visionWardsBoughtInGame: number;
    pentaKills: number;
    totalHeal: number;
    totalMinionsKilled: number;
    timeCCingOthers: number;
    altarsCaptured: number;
    altarsNeutralized: number;
    nodeCapture: number;
    nodeCaptureAssist: number;
    nodeNeutralize: number;
    nodeNeutralizeAssist: number;
    teamObjective: number;
    statPerk0: number;
    statPerk1: number;
    statPerk2: number;
}
export interface IDBXrefParticipantPerk {
    id?: number;
    participantId: number;
    perkId: string;
    varId: number;
    description?: string;
    value: number;
}
export interface IDBXrefParticipantItem {
    id?: number;
    participantId: number;
    itemId: number;
}